requirements.txt
fastapi
uvicorn[standard]
requests
faiss-cpu
pydantic
